test(research): extract PDF joining URL fixtures
